pyproject.toml
[project]
name = "financial-news-sentiment-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "boto3>=1.40.47",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "requests-html>=0.10.0",
    "torch>=2.8.0",
    "transformers>=4.57.0",
    "yahoo-fin>=0.8.9.1",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "isort>=6.1.0",
    "mypy>=1.18.2",
    "pylint>=3.3.9",
    "pytest>=8.4.2",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.0",
    "types-boto3>=1.40.46",
    "types-boto3-s3>=1.40.26",
]
model-building = [
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]
